// src/app/core/libs/scripts/libs/all-workers-api/model/candidate.ts
/**
 * Api documentation
 * Api documentation
 *
 * OpenAPI spec version: 1.0.0
 * Contact: codeguessdev237@codeguess.cm
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import {Competence} from "./competence";
import {Specialism} from "./specialism";
import {Group} from "./group";

export interface Candidate {
    qrCode?: string;
    enable?: boolean;
    isPartner?: boolean;
    /**
     * me rendre visible et disponible en ligne
     */
    publishedOnline?: boolean;
    /**
     * apres la modification des information l admintrateur doit verifier le profile si celui ci est conforme
     */
    profileUpdate: string;
    /**
     * verifier les document cni , passport
     */
    profileVerifyByAdmin?: boolean;
    profileCertificate?: boolean;
    myLogo?: string;
    picture?: string;
    email: string;
    password?: string;
    fullName: string;
    ownerName?: string;
    pseudo?: string;
    phoneNumber: string;
    gender?: string;
    dayBirth?: string;
    postBox?: string;
    rating?: number;
    dateStartExperience?: string;
    currentSalary?: number;
    city?: string;
    completeAddress?: string;
    nationality?: string;
    longitude?: number;
    latitude?: number;
    /**
     * &#039;Pending&#039;, &#039;In Progress&#039;, &#039;Completed&#039;, &#039;Cancelled&#039;
     */
    statusUser: string;
    statusReceiverNotificationJob?: boolean;
    website?: string;
    overview?: string;
    emailVerifiedAt?: Date;
    lastConnection?: Date;
    rememberToken?: string;
    linkGoogle?: string;
    linkTwitter?: string;
    linkFacebook?: string;
    linkLinkedin?: string;
    linkInstagram?: string;
    readonly createdAt?: Date;
    readonly updatedAt?: Date;
    readonly deletedAt?: Date;
    specialisms_id: number
    groupes_id: number
    tasks_count: number
    group: Group
    specialism: Specialism
    competences: Competence[]
}